requirements_dev.txt
-r requirements.txt
bump2version==1.0.1
wheel==0.41.2
django-environ==0.11.0
djangorestframework==3.14.0
pytz==2023.3
Sphinx==7.2.4
